New AsyncMockHttpServer implementation.
Requests are managed asynchronously. AsyncRequestHandler returns a
Future<HttpResponse> instead a plain HttpResponse.

This is implemented using higher kinded types, the base implementation
MockHttpServerK, and is parametrized to and F that is a Kind, and the
method execute returns a Higher1<F, HttpResponse>.

The sync version returns a Higher1<Id.µ, HttpResponse> and the async
version a Higher1<Future.µ, HttpResponse>